complete refactor of the decoder
